pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lyingdocs"
version = "0.1.0"
description = "Autonomous documentation-code misalignment detection using LLM agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "build"]

[project.scripts]
lyingdocs = "lyingdocs.cli:main"

[tool.hatch.build.targets.sdist]
include = ["lyingdocs/"]

[tool.hatch.build.targets.wheel]
packages = ["lyingdocs"]
